requirements.txt
annotated-types==0.7.0
audioop-lts==0.2.1
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comtypes==1.4.9
contourpy==1.3.1
cycler==0.12.1
Flask==3.1.0
flask-cors==6.0.0
fonttools==4.56.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
gTTS==2.5.4
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
MouseInfo==0.1.3
numpy==2.2.3
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
PyAutoGUI==0.9.54
pydantic==2.10.6
pydantic_core==2.27.2
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.2.1
pyperclip==1.9.0
pypiwin32==223
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
pyttsx3==2.98
pytweening==1.2.0
pywhatkit==5.4
pywin32==308
requests==2.32.3
rsa==4.9
six==1.17.0
soupsieve==2.6
SpeechRecognition==3.14.1
standard-aifc==3.13.0
standard-chunk==3.13.0
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
wikipedia==1.4.0
